Minor fixes.
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@1735 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ benchmarks/cmsearch-rmark/rmark_clusterfy.pl
@@ -85,6 +85,8 @@ if(! (-e ("rmark.pl"))) { die("ERROR, rmark.pl must exist in the current directo
 else { system("cp rmark.pl $run_dir"); } 
 if(! (-e ("rmark_process_glbf.pl"))) { die("ERROR, rmark_process_glbf.pl must exist in the current directory."); } 
 else { system("cp rmark_process_glbf.pl $run_dir"); } 
+if(! (-e ("rmark_times.pl"))) { die("ERROR, rmark_process_glbf.pl must exist in the current directory."); } 
+else { system("cp rmark_times.pl $run_dir"); } 
 if(! (-e ("infernal.pm"))) { die("ERROR, infernal.pm must exist in the current directory."); } 
 else { system("cp infernal.pm $run_dir"); } 
 if(! (-e ("infernal2glbf.pl"))) { die("ERROR, infernal2glbf.pl must exist in the current directory."); } 
@@ -187,7 +189,7 @@ $all_time_out = $out_file_root . "_all_time.concat";
 print PP ("cat *.glbf > $all_glbf_out\n");
 print PP ("cat *.time > $all_time_out\n");
 #11.25.05 - get timing info
-print PP ("perl ~/src/scripts/rmark/rmark_times.pl *.time > merged_" . $out_file_root . ".time\n");
+print PP ("perl rmark_times.pl *.time > merged_" . $out_file_root . ".time\n");
 
 # Call rmark_process_glbf.pl with defaults: 'hit' resolution mode and 
 # ignore cross-hits on both strands.
